+++ CHANGELOG.md
@@ -9,6 +9,7 @@ FEATURES:
 BUGS:
 
 * Fix pki config resources to allow unsetting of fields (to empty fields) ([#2558](https://github.com/hashicorp/terraform-provider-vault/pull/2558))
+* Fix tune auth mounts to allow unsetting of fields (setting fields to empty values) ([#2605](https://github.com/hashicorp/terraform-provider-vault/pull/2605))
 
 ## 5.3.0 (Sep 4, 2025)